Extract game progression logic from view.
